refactor: split four files along the seams they already had
structural/validate.go mixed schema conformance with a general LCS line
diff; diff_test.go was already separate, and now diff.go is too. triage.go
mixed the promotion workflow with 300 lines of pull-request markdown, which
moves to comment.go. gateservice.go mixed the poll loop with the
comment-embedded verdict history -- a format with a writer and a reader that
must not drift, and nothing to do with when the gate runs -- now
gatehistory.go. evals.Run dispatched two paths and inlined the third.

All four are moves. The only behaviour change is buildUserPrompt, which
declared an error it never returned, so both call sites carried dead
handling: one would have escalated, the other returned nil, for a condition
that could not occur. Underneath it, two `continue`s silently dropped a file
that would not read -- and that prompt is also the evidence string
edits.Policy.corroborated checks proposed versions against, so a silent skip
narrowed what the applier would accept, invisibly, at the moment the model
most needed the value that was dropped. The skipped paths are named now.
